+++ karbon/plugins/filtereffects/CompositeEffect.cpp
@@ -109,10 +109,10 @@ QImage CompositeEffect::processImages(const QList<QImage> &images, const KoFilte
                 da *= 255.0;
 
                 // set pre-multiplied color values on destination image
-                d = qRgba(static_cast<quint8>(qBound(0.0, dr * da, 255.0)),
-                          static_cast<quint8>(qBound(0.0, dg * da, 255.0)),
-                          static_cast<quint8>(qBound(0.0, db * da, 255.0)),
-                          static_cast<quint8>(qBound(0.0, da, 255.0)));
+                d = qRgba(static_cast<quint8>(qBound(qreal(0.0), dr * da, qreal(255.0))),
+                          static_cast<quint8>(qBound(qreal(0.0), dg * da, qreal(255.0))),
+                          static_cast<quint8>(qBound(qreal(0.0), db * da, qreal(255.0))),
+                          static_cast<quint8>(qBound(qreal(0.0), da, qreal(255.0))));
             }
         }
     } else {
